change some long long to unsigned int

+++ Device/Bagging/BagManager.cu
@@ -62,8 +62,8 @@ int *BagManager::m_pEachFeaLenEachNodeEachBag_dh = NULL;//each feature value len
 //for pinned memory; for computing indices in multiple level tree
 unsigned int *BagManager::m_pIndicesEachBag_d = NULL;	//indices for multiple level tree of each bag
 long long *BagManager::m_pNumFvalueEachNodeEachBag_d = NULL;	//the number of feature values of each (splittable?) node
-long long *BagManager::m_pFvalueStartPosEachNodeEachBag_d = NULL;//the start position of each node
-long long *BagManager::m_pEachFeaStartPosEachNodeEachBag_d = NULL;//the start position of each feature in a node
+unsigned int *BagManager::m_pFvalueStartPosEachNodeEachBag_d = NULL;//the start position of each node
+unsigned int *BagManager::m_pEachFeaStartPosEachNodeEachBag_d = NULL;//the start position of each feature in a node
 int *BagManager::m_pEachFeaLenEachNodeEachBag_d = NULL;	//the number of values of each feature in each node
 
 //memory for each individual tree
@@ -219,8 +219,8 @@ void BagManager::AllocMem()
 	//corresponding to pinned memory; for computing indices of more than one level trees
 	checkCudaErrors(cudaMalloc((void**)&m_pIndicesEachBag_d, sizeof(unsigned int) * m_numFeaValue * m_numBag));
 	checkCudaErrors(cudaMalloc((void**)&m_pNumFvalueEachNodeEachBag_d, sizeof(long long) * m_maxNumSplittable * m_numBag));
-	checkCudaErrors(cudaMalloc((void**)&m_pFvalueStartPosEachNodeEachBag_d, sizeof(long long) * m_maxNumSplittable * m_numBag));
-	checkCudaErrors(cudaMalloc((void**)&m_pEachFeaStartPosEachNodeEachBag_d, sizeof(long long) * m_maxNumSplittable * m_numBag * m_numFea));
+	checkCudaErrors(cudaMalloc((void**)&m_pFvalueStartPosEachNodeEachBag_d, sizeof(unsigned int) * m_maxNumSplittable * m_numBag));
+	checkCudaErrors(cudaMalloc((void**)&m_pEachFeaStartPosEachNodeEachBag_d, sizeof(unsigned int) * m_maxNumSplittable * m_numBag * m_numFea));
 	checkCudaErrors(cudaMalloc((void**)&m_pEachFeaLenEachNodeEachBag_d, sizeof(int) * m_maxNumSplittable * m_numBag * m_numFea));
 
 	/********** memory for splitting node ************/
