+++ Device/Memory/gbdtGPUMemManager.cu
@@ -15,7 +15,7 @@
 int *GBDTGPUMemManager::m_pDInsId = NULL;				//all the instance ids for each key-value pair
 float_point *GBDTGPUMemManager::m_pdDFeaValue = NULL; 	//all the feature values
 int *GBDTGPUMemManager::m_pDNumofKeyValue = NULL;		//the number of key-value pairs of each feature
-long long *GBDTGPUMemManager::m_pFeaStartPos = NULL;	//start key-value position of each feature
+unsigned int *GBDTGPUMemManager::m_pFeaStartPos = NULL;	//start key-value position of each feature
 //memory for instances (key on instance id)
 int *GBDTGPUMemManager::m_pDFeaId = NULL;				//all the feature ids for every instance
 float_point *GBDTGPUMemManager::m_pdDInsValue = NULL;	//all the feature values for every instance
@@ -81,7 +81,7 @@ void GBDTGPUMemManager::allocMemForIns(int nTotalNumofValue, int numofIns, int n
 	checkCudaErrors(cudaMalloc((void**)&m_pDInsId, sizeof(int) * m_totalNumofValues));
 	checkCudaErrors(cudaMalloc((void**)&m_pdDFeaValue, sizeof(float_point) * m_totalNumofValues));
 	checkCudaErrors(cudaMalloc((void**)&m_pDNumofKeyValue, sizeof(int) * m_numofFea));
-	checkCudaErrors(cudaMalloc((void**)&m_pFeaStartPos, sizeof(long long) * m_numofFea));
+	checkCudaErrors(cudaMalloc((void**)&m_pFeaStartPos, sizeof(unsigned int) * m_numofFea));
 	//memory for instances (key on instance id)
 	checkCudaErrors(cudaMalloc((void**)&m_pDFeaId, sizeof(int) * m_totalNumofValues));
 	checkCudaErrors(cudaMalloc((void**)&m_pdDInsValue, sizeof(float_point) * m_totalNumofValues));
